gui: Reorganize
