--- lara_simple/src/module-info.java
@@ -1,5 +1,0 @@
-module laraFinal {
-	
-	requires java.desktop;
-	
-}
