+++ pom.xml
@@ -27,7 +27,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>dependency/</classpathPrefix>
-                            <mainClass>application.Main</mainClass>
+                            <mainClass>org.lara.gui.Main</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
@@ -53,7 +53,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.4</version>
                 <configuration>
-                    <mainClass>application.Main</mainClass>
+                    <mainClass>org.lara.gui.Main</mainClass>
                 </configuration>
             </plugin>
         </plugins>
