+++ src/main/java/org/lara/gui/ChatFrameCtrl.java
@@ -1,4 +1,4 @@
-package application;
+package org.lara.gui;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
