--- src/main/java/org/lara/gui/bin/ChatFrame.fxml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.text.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane fx:id="anchorPane" minHeight="600.0" minWidth="1000.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.ChatFrameCtrl">
-   <children>
-      <SplitPane dividerPositions="0.29797979797979796" layoutX="175.0" layoutY="220.0" prefHeight="600.0" prefWidth="1000.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-        <items>
-          <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
-               <children>
-                  <Label fx:id="welcomeLabel" alignment="CENTER" layoutX="67.0" layoutY="56.0" prefHeight="124.0" prefWidth="159.0" text="Bienvenue, voici LARA..." wrapText="true" AnchorPane.leftAnchor="67.0" AnchorPane.rightAnchor="67.0" AnchorPane.topAnchor="56.0">
-                     <font>
-                        <Font size="21.0" />
-                     </font>
-                  </Label>
-                  <Button fx:id="buttonLouis" layoutX="82.0" layoutY="257.0" mnemonicParsing="false" onAction="#idLouis" text="Parler avec Louis" AnchorPane.topAnchor="257.0" />
-                  <Button fx:id="buttonAnna" layoutX="82.0" layoutY="318.0" mnemonicParsing="false" onAction="#idAnna" text="Parler avec Anna" AnchorPane.bottomAnchor="255.0" />
-                  <Button fx:id="buttonRiad" layoutX="84.0" layoutY="385.0" mnemonicParsing="false" onAction="#idRiad" text="Parler avec Riad" AnchorPane.bottomAnchor="190.0" />
-                  <Button fx:id="buttonAntoine" layoutX="74.0" layoutY="458.0" mnemonicParsing="false" onAction="#idAntoine" text="Parler avec Antoine" AnchorPane.bottomAnchor="115.0" />
-                  <Button fx:id="exitButton" layoutX="22.0" layoutY="553.0" mnemonicParsing="false" onAction="#exitLara" text="Quitter LARA..." AnchorPane.bottomAnchor="20.0" />
-               </children>
-            </AnchorPane>
-          <AnchorPane minHeight="0.0" minWidth="0.0" onKeyPressed="#sendKeyPressed" prefHeight="160.0" prefWidth="100.0">
-               <children>
-                  <SplitPane dividerPositions="0.8540268456375839" layoutX="101.0" layoutY="149.0" orientation="VERTICAL" prefHeight="598.0" prefWidth="697.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-                    <items>
-                      <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="100.0" prefWidth="160.0">
-                           <children>
-                              <VBox fx:id="chatBox" prefHeight="505.0" prefWidth="695.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0" />
-                           </children>
-                        </AnchorPane>
-                      <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="100.0" prefWidth="160.0">
-                           <children>
-                              <Button fx:id="sendButton" layoutX="563.0" layoutY="29.0" mnemonicParsing="false" onAction="#sendMessage" onKeyPressed="#sendKeyPressed" text="Envoyer" AnchorPane.bottomAnchor="29.0" AnchorPane.rightAnchor="60.0" />
-                              <TextField fx:id="messageInput" layoutX="32.0" layoutY="29.0" prefHeight="25.0" prefWidth="449.0" AnchorPane.bottomAnchor="29.0" AnchorPane.leftAnchor="32.0" />
-                           </children>
-                        </AnchorPane>
-                    </items>
-                  </SplitPane>
-               </children>
-            </AnchorPane>
-        </items>
-      </SplitPane>
-   </children>
-</AnchorPane>
